Fixes and Improvements.
Refactored the draw canvas method so that it's the same for the editor canvas and the play canvas.

Updated goal image. It still looks bad tbh.

Hidden the delete local data button away in a settings menu, and made it look scary.

Added a sub-title style for the second header.

Removed the editor cache, since I'm no longer resetting the data when you enter and leave the editor, it's cached by default in the editor's game state.

Bug Fixes:
========
1) you can now only place 1 player a time in the editor. Previously, you could add multiple players, but you would only see one, and only play as one. Rather than make it possible to play as multiple players, I've opted to only have one player at a time for simplicity.

2) keyboard buttons no longer do any game logic when you're not in game mode

Yet to do:
======
Make a back-to-editor button.