Updated font, dynamic color logic, background and style
Panel color alters the L of the HSL of the theme color. Font color switches between white/black and depends in the L of the HSL of the panel color. It's pretty good. Could be improved, and include button colors and floor tile color too.

Scrapped gradients. They looks good origionally, but not with my new tiling background image.