refactor. confetti. Move interruption. Colors
side effects of separating the dark overlay and the popups is that the popups look a little different now. more  vertical separation. This was done so that the confetti could appear over the dark overlay but under the popup

asset management moved out of main.js.

click moving can now be interrupted by key movement.

Confetti effect plays when you win.

Theme color now adjusts the panel colour instead of the background. Background stays a constant color. I didn't quite like the lightness adjustments I was doing before to calculate 'other' color. More work could be done on this, but at the cost of time.